# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
...
  /api/v1/watchdog/stats:
    get:
      tags: [Watchdog]
      summary: Watchdog statistics
      operationId: getWatchdogStats
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/NetworkParam"
      responses:
        "200":
          description: Watchdog summary counts
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchdogStats"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/watchdog/alerts:
    get:
      tags: [Watchdog]
      summary: List watchdog alerts
      operationId: listWatchdogAlerts
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [Info, Warning, Critical]
          description: Severity filter.
        - $ref: "#/components/parameters/NetworkParam"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
          description: Maximum alerts to return (default 100).
        - $ref: "#/components/parameters/AlertCursor"
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: object
                required: [alerts, next_cursor]
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContractAlert"
                  next_cursor:
                    type: string
                    description: Cursor for the next (older) page; empty when no more pages.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/watchdog/contracts:
    get:
      tags: [Watchdog]
      summary: List monitored contracts
      operationId: listMonitoredContracts
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
      responses:
        "200":
          description: Page of monitored contracts
          content:
            application/json:
              schema:
                type: object
                required: [contracts, next_cursor]
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: "#/components/schemas/MonitoredContract"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/watchdog/contracts/{id}:
    get:
      tags: [Watchdog]
      summary: Get one monitored contract
      operationId: getMonitoredContract
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The monitored contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoredContract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/watchdog/contracts/{id}/uptime:
    get:
      tags: [Watchdog]
      summary: Get uptime percentage for a monitored contract
      operationId: getContractUptime
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: window
          in: query
          required: false
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 24h
          description: Time window for uptime calculation (default 24h).
      responses:
        "200":
          description: Uptime result
          content:
            application/json:
              schema:
                type: object
                required: [contract_id, window, uptime_pct]
                properties:
                  contract_id:
                    type: string
                  window:
                    type: string
                  uptime_pct:
                    type: number
                    format: double
                    description: Percentage of checks with Healthy status (0–100).
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist endpoints read the user ID from X-User-ID.
  parameters:
    ContractID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Contract ID (starts with C).
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor from a previous response.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
      description: Page size (default 50, clamped to 200).
    NetworkParam:
      name: network
      in: query
      required: false
      schema:
        type: string
        enum: [testnet, mainnet, futurenet, standalone]
      description: Network filter; empty means all networks.
...
    AlertCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque alert cursor from a previous page's next_cursor.
  headers:
    XCache:
      description: |
        Response cache status (`HIT` or `MISS`). Present when the API runs
        with Redis caching enabled; cached bodies live for API_CACHE_TTL
        (default 30s) and are purged by writes to the same resource.
      schema:
        type: string
        enum: [HIT, MISS]
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidInput:
      description: Invalid query or body input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: object
              required: [code, message]
              properties:
                code:
                  type: string
                  enum: [NOT_FOUND, INVALID_INPUT, INTERNAL, RATE_LIMITED, UNSUPPORTED_MEDIA_TYPE]
                message:
                  type: string
                request_id:
                  type: string
            - type: object
              properties:
                error:
                  type: string
...
    WatchdogStats:
      type: object
      required: [total_monitored, healthy, degraded, unresponsive, total_alerts, critical_alerts]
      properties:
        total_monitored:
          type: integer
          format: int64
        healthy:
          type: integer
          format: int64
        degraded:
          type: integer
          format: int64
        unresponsive:
          type: integer
          format: int64
        total_alerts:
          type: integer
          format: int64
        critical_alerts:
          type: integer
          format: int64
    ContractAlert:
      type: object
      required: [contract_id, severity, message, ledger, tx_hash, timestamp]
      properties:
        contract_id:
          type: string
        severity:
          type: string
          enum: [Info, Warning, Critical]
        message:
          type: string
        ledger:
          type: integer
          format: int64
        tx_hash:
          type: string
        timestamp:
          type: string
          format: date-time
...
    MonitoredContract:
      type: object
      required: [contract_id, network, name, owner, status, last_check, check_interval, registered_at, updated_at]
      properties:
        contract_id:
          type: string
        network:
          type: string
        name:
          type: string
        owner:
          type: string
        status:
          type: string
        last_check:
          type: string
          format: date-time
          nullable: true
        check_interval:
          type: integer
          format: int64
        registered_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...